swapped tokenizer and tokenizer2 code

+++ processors/visualize/rw_analyzer.py
@@ -446,7 +446,8 @@ def rw_analyze(fileString, countType, tokenType, windowType, keyWord, secondKeyW
     Returns:
 
     """
-    windowSize = float(windowSize)
+
+    windowSize = int(windowSize)
     windowSizeStringLines = windowSize  # for when finding strings in window need original value
     minNumOfWindows = 10
